Refactor platform-specific functions in UnsafeRun trait.
