--- modules/core-cats/native/src/main/scala/weaver/CatsUnsafeRunPlatformCompat.scala
@@ -1,12 +1,0 @@
-package weaver
-
-import cats.effect.IO
-
-private[weaver] trait CatsUnsafeRunPlatformCompat {
-  self: CatsUnsafeRun =>
-
-  def unsafeRunSync(task: IO[Unit]): Unit = ???
-
-  def background(task: IO[Unit]): CancelToken = ???
-
-}
