+++ modules/core-cats/shared/src/main/scala/weaver/CatsUnsafeRun.scala
@@ -1,21 +1,11 @@
 package weaver
 
-import scala.concurrent.Future
-
-import cats.effect.unsafe.implicits.global
-import cats.effect.{ FiberIO, IO }
+import cats.effect.{ IO }
 
 object CatsUnsafeRun extends CatsUnsafeRun
 
 trait CatsUnsafeRun extends UnsafeRun[IO] with CatsUnsafeRunPlatformCompat {
 
-  type CancelToken = FiberIO[Unit]
-
   override implicit val parallel = IO.parallelForIO
   override implicit val effect   = IO.asyncForIO
-
-  def cancel(token: CancelToken): Unit = unsafeRunSync(token.cancel)
-
-  def unsafeRunToFuture(task: IO[Unit]): Future[Unit] = task.unsafeToFuture()
-
 }
