+++ modules/core/js-native/src/main/scala/weaver/UnsafeRunPlatform.scala
@@ -1,0 +1,7 @@
+package weaver
+
+import scala.concurrent.Future
+
+private[weaver] trait UnsafeRunPlatform[F[_]] extends EffectCompat[F] {
+  def unsafeRunToFuture(task: F[Unit]): Future[Unit]
+}
