+++ modules/core/jvm/src/main/scala/weaver/UnsafeRunPlatform.scala
@@ -1,0 +1,10 @@
+package weaver
+
+private[weaver] trait UnsafeRunPlatform[F[_]] extends EffectCompat[F] {
+
+  type CancelToken
+
+  def background(task: F[Unit]): CancelToken
+  def cancel(token: CancelToken): Unit
+  def unsafeRunSync(task: F[Unit]): Unit
+}
