+++ modules/core/shared/src/main/scala/weaver/UnsafeRun.scala
@@ -12,8 +12,7 @@ trait EffectCompat[F[_]] {
 
   private[weaver] final def sleep(duration: FiniteDuration): F[Unit] =
     effect.sleep(duration)
-  private[weaver] final def fromFuture[A](
-      thunk: => scala.concurrent.Future[A]): F[A] =
+  private[weaver] final def fromFuture[A](thunk: => Future[A]): F[A] =
     effect.fromFuture(effect.delay(thunk))
   private[weaver] final def async[A](cb: (Either[Throwable,
                                                  A] => Unit) => Unit): F[A] =
@@ -26,14 +25,4 @@ trait EffectCompat[F[_]] {
  * This is meant to delegate to library-specific constructs for running effect
  * types.
  */
-trait UnsafeRun[F[_]] extends EffectCompat[F] {
-
-  type CancelToken
-
-  def background(task: F[Unit]): CancelToken
-  def cancel(token: CancelToken): Unit
-
-  def unsafeRunSync(task: F[Unit]): Unit
-  def unsafeRunToFuture(task: F[Unit]): Future[Unit]
-
-}
+trait UnsafeRun[F[_]] extends UnsafeRunPlatform[F]
